requirements.txt
pyrogram
tgcrypto
python-dotenv
yt-dlp
aiohttp